gofmt changes and some renaming of folders
