+++ authenticator/structures/user.go
